+++ conn.h
@@ -19,6 +19,6 @@ struct conn {
 int conn_init(struct conn *, int, const uint8_t[16]);
 void conn_finish(struct conn *);
 int conn_parse_packet(struct conn *, struct recv_packet *);
-int conn_write_packet(struct conn *, const struct send_packet *);
+ssize_t conn_write_packet(struct conn *, const struct send_packet *);
 
 #endif
