+++ packet.h
@@ -27,8 +27,8 @@ struct send_packet {
 };
 void make_packet(struct send_packet *, int);
 struct send_packet *finalize_packet(struct send_packet *);
-int write_packet_data(int, const uint8_t data[], size_t len);
-int write_packet(int, const struct send_packet *);
+ssize_t write_packet_data(int, const uint8_t data[], size_t len);
+ssize_t write_packet(int, const struct send_packet *);
 void write_byte(struct send_packet *, uint8_t);
 void write_short(struct send_packet *, int16_t);
 int write_varint(struct send_packet *, int);
